Move citizen events to common as it is now being used by sols and citizen events
