+++ src/functionalTest/java/uk/gov/hmcts/divorce/common/Applicant2ApproveFT.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.divorce.citizen;
+package uk.gov.hmcts.divorce.common;
 
 import io.restassured.response.Response;
 import org.junit.jupiter.api.Test;
@@ -14,13 +14,13 @@ import static net.javacrumbs.jsonunit.core.Option.IGNORING_ARRAY_ORDER;
 import static net.javacrumbs.jsonunit.core.Option.IGNORING_EXTRA_FIELDS;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.http.HttpStatus.OK;
-import static uk.gov.hmcts.divorce.citizen.event.CitizenApplicant2Approve.APPLICANT_2_APPROVE;
+import static uk.gov.hmcts.divorce.common.event.Applicant2Approve.APPLICANT_2_APPROVE;
 import static uk.gov.hmcts.divorce.testutil.CaseDataUtil.caseData;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.ABOUT_TO_SUBMIT_URL;
 import static uk.gov.hmcts.divorce.testutil.TestResourceUtil.expectedResponse;
 
 @SpringBootTest
-public class CitizenApplicant2ApproveFT extends FunctionalTestSuite {
+public class Applicant2ApproveFT extends FunctionalTestSuite {
 
     private static final String REQUEST = "classpath:request/casedata/ccd-callback-casedata-applicant2-approved.json";
     private static final String RESPONSE = "classpath:responses/response-applicant2-approved.json";
