+++ src/main/java/uk/gov/hmcts/divorce/solicitor/event/SolicitorSubmitJointApplication.java
@@ -86,5 +86,4 @@ public class SolicitorSubmitJointApplication implements CCDConfig<CaseData, Stat
 
         return SubmittedCallbackResponse.builder().build();
     }
-
 }
